#22 - update backend / relate modules func

+++ stdmanager/src/main/java/uni/it/stdmanager/modules/vi_registration/service/impl/CourseRegistrationServiceImpl.java
@@ -156,7 +156,7 @@ public class CourseRegistrationServiceImpl implements CourseRegistrationService 
                 .id(course.getId())
                 .courseCode(course.getCourseCode())
                 .courseName(course.getCourseName())
-                .credits(course.getCredits().doubleValue())
+                .credits(course.getCredits())
                 .build();
     }
